Added Delete Function
Still not ready to Go but it should be in the next couple of days, User Schema will be Good to go for tonight, probably adding the Login Function as Well, a couple more days till the Auth. with JWT will be implemented since we will need the app to test every action first.